requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.3.3
autopep8==1.4.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
Click==7.0
command-not-found==0.3
confusable-homoglyphs==3.2.0
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap3==12.0.3
django-registration==3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
Flask==1.1.1
httplib2==0.9.2
idna==2.6
importlib-metadata==1.5.2
importlib-resources==1.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psycopg2-binary==2.8.4
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyperclip==1.7.0
pyRFC3339==1.0
python-apt==1.6.0
python-debian==0.1.32
python-decouple==3.3
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.0.14
wadllib==1.3.2
Werkzeug==1.0.0
whitenoise==5.0.1
wrapt==1.11.2
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
